Bug : La pagination ne fonctionnait plus.
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@20480 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/system/ajax_list_long_request_logs.php
@@ -12,7 +12,7 @@
 
 CCanDo::checkAdmin();
 
-$start  = CValue::get("start", 0);
+$start = (int) CValue::get("start", 0);
 
 $filter = new CLongRequestLog();
 $filter->_date_min = CValue::get("_date_min");
